updated user.js file
